requirements.txt
beautifulsoup4
defang
dnspython
lxml
python-dateutil
python-whois
selenium
tldextract
yara
